+++ tests/Fixtures/TestInvoiceLine.php
@@ -20,10 +20,11 @@ class TestInvoiceLine extends Model implements ProvidesActivityTitle
 
     public function getActivitylogOptions(): LogOptions
     {
-        return LogOptions::defaults()
-            ->logFillable()
-            ->logOnlyDirty()
-            ->dontSubmitEmptyLogs();
+        return TestLogOptions::dontLogEmpty(
+            LogOptions::defaults()
+                ->logFillable()
+                ->logOnlyDirty(),
+        );
     }
 
     public function activityTitle(): ?string